Problem/solution is now user configurable
- Also tweaked loggers to enable log restarts and/or replacement (and/or
keeping views as backups of pre-reset state); This is aimed towards
potentially supporting multiple simultaneous and/or historical solution
views, e.g. with tabbed panes.
- Also fixed an unencountered bug with PCLogger in which getting the
start time was not thread safe.
- Refactored MemeticoContentController such that it's framerate is
externally configurable.

+++ memetico/src/java/memetico/logging/PCAlgorithmState.java
@@ -34,14 +34,17 @@ public class PCAlgorithmState {
     public final String instanceName;
     public final int generation;
     public final int nAry;
-    public final double logTime;
+    /**
+     * Measured in nanoseconds
+     */
+    public final long logTime;
     public final AgentState[] agents;
 
     /**
      * Note: the cost value (in the SolutionStructure class) is expected to be already computed for all srcs
      * @param src
      */
-    public PCAlgorithmState(String instanceName, Population src, int generation, double logTime) {
+    public PCAlgorithmState(String instanceName, Population src, int generation, long logTime) {
         this.instanceName = instanceName;
         this.generation = generation;
         this.nAry = src.n_ary;
